refactor(canister-core): extract the certification layer into a Certifier
The certified-response tree, the per-rule certified entries, and the env
cookie were three fields on the State god object, and the ~13 methods that
maintain them (recertify_asset, certify_asset_encodings, range_response_certs,
on_redirect_rules_change, build_rule_entry, build_alias_rule_entry,
certify_not_found_fallback, the protection certify helpers, effective_headers)
were interleaved with serving, sync, and storage code across state.rs. The
"certified leaf must be byte-identical to the served response" invariant was
spread thin.

Move all three fields and every certify op onto a new `Certifier` type. It
takes `&Store` for the metadata/protection/token reads it needs (it's derived
state, rebuilt from the store on upgrade), and owns the env cookie — the sole
input to a certified response that isn't in the store. Serving stays on State
and reaches the tree only through read accessors (witness_to_header*,
rule_entry, root_hash, effective_headers), so it can produce witnesses but
can't mutate the certified set. State's mutation paths now write the store
then ask the Certifier to re-certify.

State's public method surface is unchanged (on_redirect_rules_change stays as
a thin delegator for sync.rs; root_hash/effective-headers callers rewired
internally), so sync.rs, lib.rs, tests.rs, and benches.rs are untouched.

Verified: builds (incl. canbench-rs), 158 unit tests pass, fmt + clippy clean,
full workspace builds, e2e 13/13 pass.
