+++ crates/canister-core/src/state.rs
@@ -1,11 +1,23 @@
-//! This module contains a pure implementation of the certified assets state machine.
+//! The certified-assets state machine: the orchestrator that ties the durable
+//! store to the certification layer and exposes the canister's behavior.
 //!
-//! Durable state (settings, per-asset metadata, content chunks) lives in stable
-//! memory via `ic-stable-structures`; derived state (the certified-response
-//! tree, per-rule certified entries) and transient upload state (staged chunks,
-//! the active sync session) live in the heap. There is no serialize/deserialize
-//! step across upgrades — `pre_upgrade` is gone and `post_upgrade` only rebuilds
-//! the derived heap state (see [`State::post_upgrade_rebuild`]).
+//! [`State`] itself holds no stable structures and no certified tree directly. It
+//! composes two owned pieces plus the small transient/heap bits that don't belong
+//! to either:
+//! - [`crate::store::Store`] — all durable state in stable memory (settings,
+//!   per-asset metadata, content chunks, tokens). The sole owner of the
+//!   `ic-stable-structures` handles and the memory layout.
+//! - [`crate::certifier::Certifier`] — the derived certified-response tree, the
+//!   per-rule certified entries, the env cookie, and the policy that maintains
+//!   them so the certified leaf always matches the served response.
+//! - transient upload state (staged chunks, the active sync session) and the
+//!   hot-path token gate index, which live on `State`.
+//!
+//! `State`'s methods are the orchestration: an asset mutation writes the `Store`
+//! then asks the `Certifier` to re-certify; serving reads the `Store` for content
+//! and the `Certifier` for witnesses. There is no serialize/deserialize step
+//! across upgrades — `pre_upgrade` is gone and `post_upgrade` only rebuilds the
+//! derived heap state from the store (see [`State::post_upgrade_rebuild`]).
 //!
 //! NB. This module does not depend on `ic_cdk` for environment access (time,
 //! certificates): those are passed in as formal arguments so the state machine
@@ -14,15 +26,11 @@
 //! in-process `VectorMemory` off-wasm (which is what lets the tests drive a full
 //! upgrade roundtrip over a shared memory handle).
 
-use crate::asset::{
-    certificate_expression_for, headers_for, range_certificate_expression_for, range_headers_for,
-    range_response_hash, response_hashes_for,
-};
+use crate::asset::{headers_for, range_headers_for};
 use crate::certification::{
-    build_ic_certificate_expression_header, response_hash, AssetKey, AssetPath,
-    CertificateExpression, CertifiedResponses, HashTreePath, NestedTreeKey, RequestHash,
-    ResponseHash,
+    build_ic_certificate_expression_header, AssetKey, AssetPath, HashTreePath,
 };
+use crate::certifier::Certifier;
 use crate::http::{HeaderField, HttpRequest, HttpResponse};
 use crate::protection::{ProtectionResponse, ProtectionStatus, TokenInfo};
 use crate::stable_store::{AssetMeta, EncodingMeta, TokenMeta};
@@ -30,8 +38,7 @@ use crate::store::Store;
 use crate::sync::{Chunk, SyncSession};
 use crate::url::url_decode;
 use candid::Principal;
-use ic_certification::{AsHashTree, Hash};
-use ic_representation_independent_hash::Value;
+use ic_certification::Hash;
 use ic_stable_structures::DefaultMemoryImpl;
 use serde_bytes::ByteBuf;
 use std::collections::{BTreeMap, HashMap};
@@ -39,7 +46,7 @@ use std::convert::TryInto;
 
 use wire_types::{
     AssetDetails, AssetEncodingDetails, CreateAssetArguments, DeleteAssetArguments, Encoding,
-    RedirectRule, RulePattern, SessionId, SetAssetContentArguments, SetAssetHeadersArguments,
+    RedirectRule, SessionId, SetAssetContentArguments, SetAssetHeadersArguments,
     UnsetAssetContentArguments,
 };
 
@@ -99,25 +106,18 @@ pub struct State {
     pub(crate) sync_session: Option<SyncSession>,
 
     // ---- derived heap (rebuilt in post_upgrade) ----
-    asset_hashes: CertifiedResponses,
-    /// Per-rule certified-tree entries, parallel to the stored redirect rules. A
-    /// `None` slot means the rule has no certified entry — either because an
-    /// asset shadows an exact rule at the same path, or because an alias rule
-    /// (200/4xx) points at a target asset that doesn't exist yet.
-    rule_certified_entries: Vec<Option<crate::redirect::CertifiedRuleEntry>>,
-    /// The fully rendered `Set-Cookie: ic_env=…` value layered onto every
-    /// `text/html` response, or `None` before any env snapshot has been
-    /// captured. Owned by the canister (never stored in `meta.headers`) and
-    /// recomputed on capture; rebuilt from the live system API in `post_upgrade`
-    /// (the env vars themselves survive as canister settings), exactly like
-    /// `asset_hashes`. See [`State::effective_headers`].
-    env_cookie: Option<String>,
+    /// The certified-response tree, the per-rule certified entries, the env
+    /// cookie, and the policy that maintains them. All derived from `store`
+    /// (plus the env snapshot) and rebuilt on upgrade. Serving reads it for
+    /// witnesses; the mutation paths drive its (re)certification. See
+    /// [`Certifier`].
+    certifier: Certifier,
     /// Hot-path gate index: `SHA-256(value) -> expires_at`. The per-request gate
     /// hashes the presented cookie and looks it up here — a heap-map read, no
     /// stable access or `TokenMeta` deserialize. Pure derived state (every entry
     /// is reconstructable from the token store, which carries each token's
     /// `token_id` and expiry), so it lives in heap and is rebuilt on upgrade like
-    /// `asset_hashes`. Kept one-to-one with the token store; revocation/expiry
+    /// the certified tree. Kept one-to-one with the token store; revocation/expiry
     /// are live.
     token_index: HashMap<[u8; 32], u64>,
 }
@@ -144,9 +144,7 @@ impl State {
             store: Store::new(memory),
             chunks: Vec::new(),
             sync_session: None,
-            asset_hashes: CertifiedResponses::default(),
-            rule_certified_entries: Vec::new(),
-            env_cookie: None,
+            certifier: Certifier::default(),
             token_index: HashMap::new(),
         }
     }
@@ -190,7 +188,7 @@ impl State {
     }
 
     pub fn root_hash(&self) -> Hash {
-        self.asset_hashes.root_hash()
+        self.certifier.root_hash()
     }
 
     // ---- asset mutations ----
@@ -303,7 +301,7 @@ impl State {
                 content_len,
             },
         );
-        self.recertify_asset(&arg.key, &meta);
+        self.certifier.recertify_asset(&self.store, &arg.key, &meta);
         self.store.put_asset(arg.key, meta);
 
         Ok(())
@@ -317,7 +315,7 @@ impl State {
 
         if let Some(old) = meta.encodings.remove(&arg.encoding) {
             self.store.delete_content_group(old.content_id);
-            self.recertify_asset(&arg.key, &meta);
+            self.certifier.recertify_asset(&self.store, &arg.key, &meta);
             self.store.put_asset(arg.key, meta);
         }
 
@@ -326,7 +324,7 @@ impl State {
 
     pub fn delete_asset(&mut self, arg: DeleteAssetArguments) {
         if let Some(meta) = self.store.remove_asset(&arg.key) {
-            self.asset_hashes.remove_responses_for_path(&arg.key);
+            self.certifier.remove_responses_for_path(&arg.key);
             for enc in meta.encodings.values() {
                 self.store.delete_content_group(enc.content_id);
             }
@@ -340,39 +338,10 @@ impl State {
             .ok_or_else(|| "asset not found".to_string())?;
 
         meta.headers = arg.headers;
-        self.recertify_asset(&arg.key, &meta);
+        self.certifier.recertify_asset(&self.store, &arg.key, &meta);
         self.store.put_asset(arg.key, meta);
 
         Ok(())
-    }
-
-    /// The header list a response actually certifies and serves: the asset's own
-    /// `meta.headers`, plus the canister-owned env cookie on `text/html` assets
-    /// when a snapshot exists. The cookie is *never* stored in `meta.headers`
-    /// (`_headers` remains the sole owner of that field); it is layered on here
-    /// at the single point every cert/serve site funnels through, guaranteeing
-    /// the certified set and the served set agree. A user's own `_headers`
-    /// `Set-Cookie` coexists as a separate entry (per-asset headers are a `Vec`,
-    /// not a name-keyed map, so neither overwrites the other).
-    fn effective_headers(&self, meta: &AssetMeta) -> Vec<(String, String)> {
-        let mut headers = meta.headers.clone();
-        if let Some(cookie) = &self.env_cookie {
-            if crate::asset::is_html_content_type(&meta.content_type) {
-                headers.push(("set-cookie".to_string(), cookie.clone()));
-            }
-        }
-        // Under protection, force `Cache-Control: no-store` on every served
-        // response. The boundary cache is cookie-blind (keys on path+range, no
-        // `Vary`), so a cached `200` could be replayed to a no-token request
-        // (asset leak) or a cached `307`/`401` to an authorized user. `no-store`
-        // is certified like any other header, so an honest gateway can't strip
-        // it. This overrides any user `_headers` cache-control while protected;
-        // public apps (`protection == None`) are untouched.
-        if self.protection_enabled() {
-            headers.retain(|(k, _)| !k.eq_ignore_ascii_case("cache-control"));
-            headers.push(("cache-control".to_string(), "no-store".to_string()));
-        }
-        headers
     }
 
     // ---- access protection ----
@@ -381,123 +350,6 @@ impl State {
     /// the in-memory `StableCell`, so it is cheap enough to call per request.
     pub fn protection_login_page(&self) -> Option<String> {
         self.store.protection_login_page()
-    }
-
-    /// Whether access protection is enabled (the gate is on).
-    pub fn protection_enabled(&self) -> bool {
-        self.store.protection_enabled()
-    }
-
-    /// Removes and recomputes every certified response for one asset from its
-    /// metadata. Recompute is cheap: it reads only headers/content_type/sha256,
-    /// never content bytes.
-    fn recertify_asset(&mut self, key: &AssetKey, meta: &AssetMeta) {
-        self.asset_hashes.remove_responses_for_path(key);
-        if !meta.encodings.is_empty() {
-            self.certify_asset_encodings(key, meta);
-        }
-
-        // Access protection: this path may also carry the gate's certified
-        // unauthenticated sibling — or, for the login page itself, its redeem
-        // responses (which share this subtree and were wiped above).
-        if let Some(login_page) = self.protection_login_page() {
-            if key.as_str() == login_page {
-                self.reassert_login_responses(&login_page);
-            } else {
-                self.certify_unauth_sibling(key, meta, &login_page);
-            }
-        }
-    }
-
-    /// Certifies the 200/304 (single-chunk) or N×206 (multi-chunk) responses for
-    /// every encoding of an asset. Split out of [`Self::recertify_asset`] so the
-    /// access-protection tail there runs even for a content-less asset.
-    fn certify_asset_encodings(&mut self, key: &AssetKey, meta: &AssetMeta) {
-        let effective_headers = self.effective_headers(meta);
-        let path = AssetPath::from(key.as_str());
-        for (&encoding, enc) in &meta.encodings {
-            let cert_expr = certificate_expression_for(&effective_headers, encoding, &enc.sha256);
-            let response_hashes = response_hashes_for(
-                &effective_headers,
-                &meta.content_type,
-                encoding,
-                &cert_expr,
-                &enc.sha256,
-            );
-            // Always certify the 304 (empty body). Certify the full 200 only for
-            // single-chunk encodings — a multi-chunk asset is served as N×206 (the
-            // gateway reassembles them into a 200), so a full 200 is never served
-            // and certifying it would be dead weight in the tree.
-            let hash_304 = path.hash_tree_path(
-                &cert_expr,
-                &RequestHash::default(),
-                ResponseHash::from(&response_hashes[&304]),
-            );
-            self.asset_hashes.certify_response_precomputed(&hash_304);
-            if enc.num_chunks == 1 {
-                let hash_200 = path.hash_tree_path(
-                    &cert_expr,
-                    &RequestHash::default(),
-                    ResponseHash::from(&response_hashes[&200]),
-                );
-                self.asset_hashes.certify_response_precomputed(&hash_200);
-            } else {
-                // Multi-chunk: certify one 206 per chunk (response-only).
-                let (range_cert_expr, resp_hashes) = self.range_response_certs(
-                    &effective_headers,
-                    &meta.content_type,
-                    encoding,
-                    enc,
-                );
-                for resp_hash in resp_hashes {
-                    let hash_path = path.hash_tree_path(
-                        &range_cert_expr,
-                        &RequestHash::default(),
-                        ResponseHash::from(&resp_hash),
-                    );
-                    self.asset_hashes.certify_response_precomputed(&hash_path);
-                }
-            }
-        }
-    }
-
-    /// Content-free per-chunk 206 certification data for a multi-chunk encoding:
-    /// the shared range certificate expression plus one response hash per chunk,
-    /// in chunk order. Derived from `chunk_certs` + `content_len`, so it never
-    /// reads chunk bytes. Both the direct-asset path (`recertify_asset`) and the
-    /// alias path (`build_alias_rule_entry`) use it, each placing the resulting
-    /// leaves at its own tree location.
-    fn range_response_certs(
-        &self,
-        effective_headers: &[(String, String)],
-        content_type: &str,
-        encoding: Encoding,
-        enc: &EncodingMeta,
-    ) -> (CertificateExpression, Vec<[u8; 32]>) {
-        let range_cert_expr =
-            range_certificate_expression_for(effective_headers, encoding, &enc.sha256);
-        let total = enc.content_len;
-        let chunk_infos: Vec<(u32, [u8; 32])> = self
-            .store
-            .chunk_certs_of(enc.content_id)
-            .map(|(_, cc)| (cc.len, cc.sha256))
-            .collect();
-        let mut resp_hashes = Vec::with_capacity(chunk_infos.len());
-        let mut offset: u64 = 0;
-        for (len, sha256) in chunk_infos {
-            let len = len as u64;
-            let content_range = format!("bytes {}-{}/{}", offset, offset + len - 1, total);
-            resp_hashes.push(range_response_hash(
-                effective_headers,
-                content_type,
-                encoding,
-                &enc.sha256,
-                &content_range,
-                &sha256,
-            ));
-            offset += len;
-        }
-        (range_cert_expr, resp_hashes)
     }
 
     // ---- queries ----
@@ -671,10 +523,7 @@ impl State {
                 if !crate::redirect::matches(rule, path) {
                     return None;
                 }
-                let entry = self
-                    .rule_certified_entries
-                    .get(idx)
-                    .and_then(|e| e.as_ref())?;
+                let entry = self.certifier.rule_entry(idx)?;
                 Some((rule, entry))
             })
     }
@@ -691,7 +540,7 @@ impl State {
         // serve) yields `None` here and falls through to the rule scan, so a
         // wildcard rule can still cover it.
         if let Some(meta) = self.store.get_asset(&path.to_string()) {
-            let (cert_header, _) = self.asset_hashes.witness_to_header(path, certificate);
+            let cert_header = self.certifier.witness_to_header(path, certificate);
             if let Some(response) = self.build_asset_response(
                 &meta,
                 &requested_encodings,
@@ -717,7 +566,7 @@ impl State {
             );
         }
 
-        let (certificate_header, _) = self.asset_hashes.witness_to_header(path, certificate);
+        let certificate_header = self.certifier.witness_to_header(path, certificate);
         HttpResponse::build_404(certificate_header)
     }
 
@@ -800,7 +649,7 @@ impl State {
         }
 
         let mut headers = headers_for(
-            &self.effective_headers(meta),
+            &self.certifier.effective_headers(&self.store, meta),
             &meta.content_type,
             encoding,
             &enc.sha256,
@@ -873,7 +722,7 @@ impl State {
         let chunk = self.store.read_chunk(enc.content_id, chunk_index);
         let content_range = format!("bytes {}-{}/{}", chunk_start, chunk_start + len - 1, total);
         let mut headers = range_headers_for(
-            &self.effective_headers(meta),
+            &self.certifier.effective_headers(&self.store, meta),
             &meta.content_type,
             encoding,
             &enc.sha256,
@@ -902,7 +751,7 @@ impl State {
         range_start: Option<usize>,
     ) -> HttpResponse {
         let cert_header =
-            self.asset_hashes
+            self.certifier
                 .witness_to_header_with_location(path, &entry.location, certificate);
         match &entry.kind {
             crate::redirect::CertifiedRuleEntryKind::Synthetic { expression } => {
@@ -940,21 +789,6 @@ impl State {
 
     // ---- access-protection certify/serve helpers ----
 
-    /// Certifies one [`ProtectionResponse`] as a response-only leaf at `location`
-    /// (an asset `<$>`, a rule `<$>`/`<*>`, or the root `<*>`). Returns the leaf's
-    /// `HashTreePath` so per-token redeem leaves can be stored for later removal.
-    fn certify_protection_response(
-        &mut self,
-        location: &HashTreePath,
-        resp: &ProtectionResponse,
-    ) -> HashTreePath {
-        let cert_expr = resp.cert_expr();
-        let resp_hash = resp.response_hash();
-        let tp = crate::redirect::alias_tree_path(location, cert_expr.expression_hash, resp_hash);
-        self.asset_hashes.certify_response_precomputed(&tp);
-        tp
-    }
-
     /// Builds the served `HttpResponse` for a [`ProtectionResponse`], attaching the
     /// `IC-CertificateExpression` header and the `IC-Certificate` witness for
     /// `location`. The bytes match the certified leaf exactly (same constructor),
@@ -969,7 +803,7 @@ impl State {
         let cert_expr = resp.cert_expr();
         let cert_expr_header = build_ic_certificate_expression_header(&cert_expr);
         let cert_header =
-            self.asset_hashes
+            self.certifier
                 .witness_to_header_with_location(request_path, location, certificate);
         let mut headers = resp.headers.clone();
         headers.push(cert_expr_header);
@@ -979,39 +813,6 @@ impl State {
             headers,
             body: ByteBuf::from(resp.body.clone()),
             upgrade: None,
-        }
-    }
-
-    /// Certifies the unauthenticated sibling for a protected asset path: a
-    /// `307 → <login_page>` for HTML (humans navigate to documents) or a `401`
-    /// for anything else (a redirect would hand a subresource the wrong content
-    /// type).
-    fn certify_unauth_sibling(&mut self, key: &AssetKey, meta: &AssetMeta, login_page: &str) {
-        let resp = if crate::asset::is_html_content_type(&meta.content_type) {
-            ProtectionResponse::redirect_to_login(login_page)
-        } else {
-            ProtectionResponse::unauthorized()
-        };
-        let location = AssetPath::from(key.as_str()).asset_hash_path_root();
-        self.certify_protection_response(&location, &resp);
-    }
-
-    /// (Re)certifies the login page's `POST` responses: the shared redeem-failure
-    /// `401` plus each live token's `302 + Set-Cookie` redeem response. The `200`
-    /// page comes from the sync at the same subtree, so this must run whenever
-    /// that subtree is rebuilt (and on upgrade). Token redeem leaves are
-    /// re-inserted from their stored paths — no plaintext needed.
-    fn reassert_login_responses(&mut self, login_page: &str) {
-        let location = AssetPath::from(login_page).asset_hash_path_root();
-        self.certify_protection_response(&location, &ProtectionResponse::redeem_failure());
-        let paths: Vec<Vec<NestedTreeKey>> = self
-            .store
-            .iter_tokens()
-            .map(|(_, meta)| meta.redeem_path)
-            .collect();
-        for path in paths {
-            self.asset_hashes
-                .certify_response_precomputed(&HashTreePath(path));
         }
     }
 
@@ -1087,47 +888,7 @@ impl State {
         self.serve_protection_response(&resp, login_page, &location, certificate)
     }
 
-    /// Adds the protection layer's redirect-rule siblings after a rule rebuild:
-    /// the universal root `<*>` `307 → login`, plus a `307` at every non-root rule
-    /// location (whose exact entry would otherwise reject the root wildcard). The
-    /// per-rule siblings are tracked on the rule entry so the next rebuild removes
-    /// them; the root one is re-added each rebuild (cleared by
-    /// `remove_fallback_responses`). Only called while protection is on.
-    fn certify_rule_unauth_siblings(&mut self, login_page: &str) {
-        let redirect = ProtectionResponse::redirect_to_login(login_page);
-        let root = HashTreePath::not_found_base_path();
-        self.certify_protection_response(&root, &redirect);
-        for idx in 0..self.rule_certified_entries.len() {
-            let Some(location) = self
-                .rule_certified_entries
-                .get(idx)
-                .and_then(|e| e.as_ref())
-                .map(|e| e.location.clone())
-            else {
-                continue;
-            };
-            if location.0 == root.0 {
-                continue; // root <*> already covered by the universal sibling
-            }
-            let tp = self.certify_protection_response(&location, &redirect);
-            if let Some(entry) = self.rule_certified_entries[idx].as_mut() {
-                entry.tree_paths.push(tp);
-            }
-        }
-    }
-
     // ---- access-protection state management (controller-driven) ----
-
-    /// Re-certifies every asset from its metadata. Shared by the protection
-    /// toggles and `post_upgrade_rebuild`.
-    fn recertify_all_assets(&mut self) {
-        let keys = self.store.asset_keys();
-        for key in keys {
-            if let Some(meta) = self.store.get_asset(&key) {
-                self.recertify_asset(&key, &meta);
-            }
-        }
-    }
 
     /// Removes every certified response under the login page's subtree (redeem
     /// `302`s, the `401`, and the `200` page if synced) and drops all tokens. A
@@ -1136,7 +897,7 @@ impl State {
         // The whole login subtree (redeem 302s, the 401, the 200 page) goes in one
         // shot, so per-token cert removal isn't needed — just empty the store and
         // the (heap) gate index.
-        self.asset_hashes.remove_responses_for_path(login_page);
+        self.certifier.remove_responses_for_path(login_page);
         self.store.clear_tokens();
         self.token_index.clear();
     }
@@ -1157,12 +918,13 @@ impl State {
         }
         self.store
             .set_protection_login_page(Some(login_page.clone()));
-        self.recertify_all_assets();
+        self.certifier.recertify_all_assets(&self.store);
         self.on_redirect_rules_change();
         // Enable-first: if the page isn't synced yet, the asset loop didn't touch
         // it, but the redeem endpoint must still work — assert its responses.
         if !self.store.contains_asset(&login_page) {
-            self.reassert_login_responses(&login_page);
+            self.certifier
+                .reassert_login_responses(&self.store, &login_page);
         }
     }
 
@@ -1174,7 +936,7 @@ impl State {
         };
         self.store.set_protection_login_page(None);
         self.clear_tokens_and_login_responses(&login_page);
-        self.recertify_all_assets();
+        self.certifier.recertify_all_assets(&self.store);
         self.on_redirect_rules_change();
     }
 
@@ -1201,6 +963,7 @@ impl State {
         let expires_at = now.saturating_add((ttl_secs as u64).saturating_mul(1_000_000_000));
         let location = AssetPath::from(login_page.as_str()).asset_hash_path_root();
         let redeem_tp = self
+            .certifier
             .certify_protection_response(&location, &ProtectionResponse::redeem_success(&value));
         let token_id = crate::protection::token_id(&value);
         self.token_index.insert(token_id, expires_at);
@@ -1222,8 +985,8 @@ impl State {
     fn remove_token(&mut self, label: &str) {
         if let Some(meta) = self.store.remove_token(label) {
             self.token_index.remove(&meta.token_id);
-            self.asset_hashes
-                .remove_response_precomputed(&HashTreePath(meta.redeem_path));
+            self.certifier
+                .remove_response(&HashTreePath(meta.redeem_path));
         }
     }
 
@@ -1331,190 +1094,20 @@ impl State {
 
     // ---- redirect rules ----
 
-    /// Rebuild the certified-tree entries for the redirect rules. Called
-    /// whenever the rule list changes (in `execute_operations`) or assets are
-    /// restored from stable memory (in `post_upgrade_rebuild`). Caller must
-    /// refresh `certified_data` after this.
-    ///
-    /// Status-200 rules borrow each encoding's certificate expression and
-    /// response hash from the target asset, so this also has to run after any
-    /// asset change that could affect those values.
-    ///
-    /// This is also the single owner of the root `<*>` fallback slot: a root
-    /// `/*` rule and the built-in 404 are mutually exclusive occupants of it, so
-    /// after rebuilding the rules we (re)certify the built-in 404 exactly when
-    /// no active rule claims `<*>`. That keeps `build_http_response`'s
-    /// fallthrough certified — the gateway rejects uncertified responses, so an
-    /// uncertified fallback 404 would be unservable.
+    /// Rebuilds the certified-tree entries for the redirect rules and the
+    /// built-in 404 fallback (and, under protection, the rule/root unauthenticated
+    /// siblings). Called whenever the rule list changes (in `execute_operations`),
+    /// after asset ops that could clobber a rule's tree slot, and on upgrade.
+    /// Delegates to the [`Certifier`]; the caller refreshes `certified_data`
+    /// afterwards.
     pub fn on_redirect_rules_change(&mut self) {
-        for entry in self.rule_certified_entries.drain(..).flatten() {
-            for tp in &entry.tree_paths {
-                self.asset_hashes.remove_response_precomputed(tp);
-            }
-        }
-        // Drop any built-in 404 before rebuilding so a rule taking over `<*>`
-        // doesn't leave a stale fallback hash beside it (and removing it now is
-        // safe — we re-add below if `<*>` ends up rule-free).
-        self.asset_hashes.remove_fallback_responses();
-
-        let rules = self.store.redirect_rules().to_vec();
-        let mut new_entries: Vec<Option<crate::redirect::CertifiedRuleEntry>> =
-            Vec::with_capacity(rules.len());
-        for rule in &rules {
-            new_entries.push(self.build_rule_entry(rule));
-        }
-        self.rule_certified_entries = new_entries;
-
-        if !self.asset_hashes.has_fallback_response() {
-            self.certify_not_found_fallback();
-        }
-
-        // Under protection, every certified path needs an unauthenticated sibling
-        // (and a universal root `<*>` fallback) so the gate can serve a verifiable
-        // 307/401 there. Added after the rules + built-in 404 so it layers on top.
-        if let Some(login_page) = self.protection_login_page() {
-            self.certify_rule_unauth_siblings(&login_page);
-        }
-    }
-
-    /// Certifies the canister's built-in last-resort 404 ("not found") at the
-    /// root `<*>` fallback path. Callers must ensure `<*>` is otherwise free;
-    /// `on_redirect_rules_change` is the only caller and gates on that.
-    fn certify_not_found_fallback(&mut self) {
-        let response = HttpResponse::uncertified_404();
-        let headers: Vec<_> = response
-            .headers
-            .into_iter()
-            .map(|(k, v)| (k, Value::String(v)))
-            .collect();
-        self.asset_hashes.certify_fallback_response(
-            response.status_code,
-            &headers,
-            &response.body,
-            None,
-        );
+        self.certifier.on_redirect_rules_change(&self.store);
     }
 
     /// Replaces the redirect rules and rebuilds their certified entries.
     pub fn set_redirect_rules(&mut self, rules: Vec<RedirectRule>) {
         self.store.set_redirect_rules(rules);
         self.on_redirect_rules_change();
-    }
-
-    fn build_rule_entry(
-        &mut self,
-        rule: &RedirectRule,
-    ) -> Option<crate::redirect::CertifiedRuleEntry> {
-        if let RulePattern::Exact(src) = &rule.from {
-            if self.store.contains_asset(src) {
-                // Asset at the source path shadows the rule.
-                return None;
-            }
-        }
-        match rule.status {
-            // 200 rewrites and 4xx custom error pages both borrow body + headers
-            // from the target asset (4xx re-certifies with the override status;
-            // see `build_alias_rule_entry`).
-            200 | 404 | 410 => self.build_alias_rule_entry(rule, rule.status),
-            // 3xx redirects synthesize an empty body; only the headers
-            // (content-type, Location) are certified.
-            _ => {
-                let entry = crate::redirect::build_synthetic_entry(rule);
-                for tp in &entry.tree_paths {
-                    self.asset_hashes.certify_response_precomputed(tp);
-                }
-                Some(entry)
-            }
-        }
-    }
-
-    fn build_alias_rule_entry(
-        &mut self,
-        rule: &RedirectRule,
-        status: u16,
-    ) -> Option<crate::redirect::CertifiedRuleEntry> {
-        let target_key = rule.to.clone();
-        let meta = self.store.get_asset(&target_key)?;
-        // Mirror the target asset's effective headers (incl. the env cookie on
-        // an HTML target) so the alias reuses the same certified response the
-        // direct hit / serve path produces.
-        let effective_headers = self.effective_headers(&meta);
-        let location = crate::redirect::tree_location(rule);
-        let mut tree_paths = Vec::new();
-        for (&encoding, enc) in &meta.encodings {
-            // A 4xx custom error page is served as a single inline body with the
-            // override status — 206 reassembly always yields a 200, so a
-            // multi-chunk encoding can't carry it. Mirror the serve-side
-            // `acceptable` filter and skip such encodings; certifying them here
-            // would leave an unservable leaf the serve path never reproduces (it
-            // falls back to the built-in 404), so the gateway would reject the
-            // mismatched response. If *every* encoding is multi-chunk the rule
-            // gets no leaves and goes inert (built-in 404 fallthrough). The sync
-            // op guard rejects this combination up front; this keeps the
-            // certified tree sound even if some other caller slips it through.
-            if status != 200 && enc.num_chunks > 1 {
-                continue;
-            }
-            let cert_expr = certificate_expression_for(&effective_headers, encoding, &enc.sha256);
-
-            // A 200-rewrite to a multi-chunk target serves N×206 (the gateway
-            // reassembles them into a 200), exactly like a direct hit. Certify
-            // those 206 leaves at the alias location and move on.
-            if status == 200 && enc.num_chunks > 1 {
-                let (range_cert_expr, resp_hashes) = self.range_response_certs(
-                    &effective_headers,
-                    &meta.content_type,
-                    encoding,
-                    enc,
-                );
-                for resp_hash in resp_hashes {
-                    let tp = crate::redirect::alias_tree_path(
-                        &location,
-                        range_cert_expr.expression_hash,
-                        resp_hash,
-                    );
-                    self.asset_hashes.certify_response_precomputed(&tp);
-                    tree_paths.push(tp);
-                }
-                continue;
-            }
-
-            let resp_hash = if status == 200 {
-                // The encoding's already-certified 200 response hash.
-                response_hashes_for(
-                    &effective_headers,
-                    &meta.content_type,
-                    encoding,
-                    &cert_expr,
-                    &enc.sha256,
-                )[&200]
-            } else {
-                // 4xx custom error page: re-certify with the override status
-                // using the same headers and body the asset would serve at 200.
-                let base_headers: Vec<(String, Value)> = headers_for(
-                    &effective_headers,
-                    &meta.content_type,
-                    encoding,
-                    &enc.sha256,
-                )
-                .into_iter()
-                .map(|(k, v)| (k, Value::String(v)))
-                .collect();
-                response_hash(&base_headers, status, &enc.sha256).0
-            };
-            let tp =
-                crate::redirect::alias_tree_path(&location, cert_expr.expression_hash, resp_hash);
-            self.asset_hashes.certify_response_precomputed(&tp);
-            tree_paths.push(tp);
-        }
-        if tree_paths.is_empty() {
-            return None;
-        }
-        Some(crate::redirect::CertifiedRuleEntry {
-            tree_paths,
-            location,
-            kind: crate::redirect::CertifiedRuleEntryKind::AliasOf { target_key, status },
-        })
     }
 
     // ---- environment cookie ----
@@ -1524,7 +1117,7 @@ impl State {
     /// the rebuild — which re-certifies every asset from scratch — picks the
     /// cookie up through `effective_headers`.
     pub fn store_env(&mut self, env: &crate::runtime::CanisterEnv) {
-        self.env_cookie = Some(env.render_cookie());
+        self.certifier.set_env_cookie(env.render_cookie());
     }
 
     /// Captures the env at the **start of a sync**, so the operations that follow
@@ -1537,7 +1130,7 @@ impl State {
     /// certificate while `effective_headers` serves the new cookie, and the
     /// gateway would reject them. Caller publishes `certified_data` afterwards.
     pub fn capture_env_at_sync_start(&mut self, env: &crate::runtime::CanisterEnv) {
-        if self.env_cookie.as_deref() != Some(env.render_cookie().as_str()) {
+        if self.certifier.env_cookie() != Some(env.render_cookie().as_str()) {
             self.refresh_env(env);
         }
     }
@@ -1554,7 +1147,7 @@ impl State {
         for key in keys {
             if let Some(meta) = self.store.get_asset(&key) {
                 if crate::asset::is_html_content_type(&meta.content_type) {
-                    self.recertify_asset(&key, &meta);
+                    self.certifier.recertify_asset(&self.store, &key, &meta);
                 }
             }
         }
@@ -1571,7 +1164,7 @@ impl State {
     /// via `on_redirect_rules_change`), and the token gate index. The caller
     /// publishes `certified_data` afterwards.
     pub fn post_upgrade_rebuild(&mut self) {
-        self.recertify_all_assets();
+        self.certifier.recertify_all_assets(&self.store);
         self.on_redirect_rules_change();
         self.rebuild_token_index();
         // If protection is on but the login page asset isn't present, the asset
@@ -1579,7 +1172,8 @@ impl State {
         // the redeem endpoint survives the upgrade even in the degraded state.
         if let Some(login_page) = self.protection_login_page() {
             if !self.store.contains_asset(&login_page) {
-                self.reassert_login_responses(&login_page);
+                self.certifier
+                    .reassert_login_responses(&self.store, &login_page);
             }
         }
     }
